finalized version of auto lab
corrected an issue that required adding in a
new chip router to add two more instances for
the lights.

This was because there were not enough available
pins when bluetooth was fully enabled to work with
the app for all 6 lighting conditions.

Reconfigured the latch pin and light pin locations
to accomadate for the new chip bypass and to match
what the app's lighting condition settings reflect.